backend/requirements.txt
python-dotenv
flask
flask-cors
requests
sentence-transformers
numpy
msgpack
torch
scipy
flasgger
